# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HiraNXT Jewellery ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repairs:
    get:
      operationId: listRepairs
      tags: [repairs]
      summary: List all repair jobs
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Repair jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepairJob"
    post:
      operationId: createRepair
      tags: [repairs]
      summary: Log a new repair job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairJobInput"
      responses:
        "201":
          description: Created repair job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairJob"
  /repairs/{id}:
    get:
      operationId: getRepair
      tags: [repairs]
      summary: Get a repair job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repair job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairJob"
    patch:
      operationId: updateRepair
      tags: [repairs]
      summary: Update a repair job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairJobUpdate"
      responses:
        "200":
          description: Updated repair job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairJob"
    delete:
      operationId: deleteRepair
      tags: [repairs]
      summary: Delete a repair job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- CUSTOM ORDERS ----------
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- SETTINGS ----------
...
components:
  schemas:
...
    RepairJob:
      type: object
      required: [id, customerId, customerName, customerMobile, itemDescription, issue, estimatedCost, status, receivedDate, promisedDate, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerMobile:
          type: string
        itemDescription:
          type: string
        issue:
          type: string
        estimatedCost:
          type: number
        actualCost:
          type: ["number", "null"]
        status:
          type: string
          description: received, in_progress, ready, delivered
        receivedDate:
          type: string
        promisedDate:
          type: string
        deliveredDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        karigarId:
          type: ["integer", "null"]
        karigarName:
          type: ["string", "null"]
        createdAt:
          type: string
    RepairJobInput:
      type: object
      required: [customerName, customerMobile, itemDescription, issue, estimatedCost, promisedDate]
      properties:
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerMobile:
          type: string
        itemDescription:
          type: string
        issue:
          type: string
        estimatedCost:
          type: number
        promisedDate:
          type: string
        notes:
          type: ["string", "null"]
        karigarId:
          type: ["integer", "null"]
    RepairJobUpdate:
      type: object
      properties:
        status:
          type: string
        actualCost:
          type: ["number", "null"]
        deliveredDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        customerName:
          type: string
        customerMobile:
          type: string
        itemDescription:
          type: string
        issue:
          type: string
        estimatedCost:
          type: number
        promisedDate:
          type: string
        karigarId:
          type: ["integer", "null"]
...
    Supplier:
      type: object
      required: [id, name, mobile, address, gstin, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        address:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        openingBalance:
          type: number
        openingBalanceType:
          type: string
          description: debit or credit — debit means you paid the supplier in advance, credit means you still owe them
        createdAt:
          type: string
    SupplierInput:
      type: object
      required: [name, mobile]
      properties:
        name:
          type: string
        mobile:
          type: string
        address:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        openingBalance:
          type: number
        openingBalanceType:
          type: string
...